Don't limit hist to numeric types
The stricture against hist tables on categorical data is
harmful. The docs explanation that no intuitive intervals exist
for non-numeric fields is wrong, since non-numeric fields are just
their own intervals. e.g. "A","A","A","B","B","C" has intervals
of "A", "B", and "C" with counts of 3, 2, and 1 accordingly.